Fixed Form passwords layout (desktop)
